requirements.txt
python-telegram-bot==20.3
Flask==2.3.3
aiohttp==3.9.3
requests==2.31.0
